refactor: 모달 타입 분리
