+++ components/src/lib/Modal.tsx
@@ -11,46 +11,14 @@ import {
   ModalWrapperStyle,
 } from "./styles";
 import { useFocusTrap } from "./hooks/useFocusTrap";
-
-export interface ChildrenProps {
-  /** 자식 요소 (JSX.Element 또는 문자열 등) */
-  children: React.ReactNode;
-}
-
-export interface ModalProps extends ChildrenProps {
-  /** 모달을 보여줄지 여부 */
-  show: boolean;
-
-  /** 모달을 닫는 함수 (배경 클릭이나 X 버튼 클릭 시 호출) */
-  onHide: () => void;
-
-  /** 배경 어두움 여부  */
-  background?: boolean;
-
-  /** 모달 위치 설정: 가운데(center), 상단(top), 하단(bottom) */
-  position?: "center" | "top" | "bottom";
-
-  /**모달의 넓이를 설정 */
-  size?: "small" | "medium" | "large";
-}
-
-export interface ModalHeaderProps extends ChildrenProps {
-  /** 닫기 버튼(X)을 표시할지 여부 */
-  closeButton?: boolean;
-}
-
-export interface ModalTitleProps extends ChildrenProps {
-  /** 텍스트 색상 설정 */
-  color?: string;
-}
-
-export interface ModalFooterProps extends ChildrenProps {
-  buttonAlign?: "left" | "center" | "right";
-}
-
-interface ModalContextType {
-  onHide: () => void;
-}
+import {
+  ChildrenProps,
+  ModalContextType,
+  ModalFooterProps,
+  ModalHeaderProps,
+  ModalProps,
+  ModalTitleProps,
+} from "./types/Modal.types";
 
 const ModalContext = createContext<ModalContextType>({
   onHide: () => {
@@ -58,14 +26,7 @@ const ModalContext = createContext<ModalContextType>({
   },
 });
 
-const Modal = ({
-  show,
-  onHide,
-  background = true,
-  position = "center",
-  size = "medium",
-  children,
-}: ModalProps) => {
+const Modal = ({ show, onHide, background = true, position = "center", size = "medium", children }: ModalProps) => {
   const containerRef = useRef<HTMLDivElement>(null);
 
   useFocusTrap(containerRef, show);
